Cairo.serial {Cairo}	R Documentation

Check for changes in the graphics state of Cairo devices.

Description

Cairo.serial returns an integer that is increased with every plotting operation on the device. This allows user code to determine whether any new content has been added to the device since it was last checked.

Usage

Cairo.serial(device = dev.cur())

Arguments

device
device number or an object of the class Cairo (as obtained from the Cairo function).

Value

Integer value.

Note

The integer value overflows to 0 at 2^31. Typically only equality should be checked and for such it is extremely unlikely that the state has changed yet the serial value is the same due to overflow.

Author(s)

Simon Urbanek

[Package Cairo version 1.5-12.2 Index]