Cairo.capabilities {Cairo}	R Documentation

Reports which output types are supported by this Cairo build

Description

Cairo.capabilities returns a logical vector describing the capabilities of this particular Cairo build.

Usage

Cairo.capabilities()

Details

The Cairo package provides multiple back-ends, such as images (PNG, JPEG, TIFF), vector graphics (PDF, PostScript, SVG) or displays (X11, Windows). However, not all systems support all back-ends. The Cairo.capabilities function returns a logical vector showing which capabilities are supported in this particular Cairo build.

Note that the capabilities depend both on the libraries available in the system as well as the compiled-in modules in cairo graphics.

See Also

Cairo

[Package Cairo version 1.4-5 Index]